1sem/LR5/requirements.txt
grpcio==1.75.1
protobuf==6.33.0rc2
flask==2.3.3